Moved django filter tests
